% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_column_editing.R
\name{birdnet_drop_datetime}
\alias{birdnet_drop_datetime}
\title{Drop datetime-related columns from BirdNET output}
\usage{
birdnet_drop_datetime(data)
}
\arguments{
\item{data}{A data frame containing BirdNET output, with the datetime related columns present.}
}
\value{
A data frame with datime related columns removed.
}
\description{
Drop datetime-related columns from BirdNET output
}
\keyword{internal}
