% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_row_editing.R
\name{birdnet_filter_hour}
\alias{birdnet_filter_hour}
\title{Filter data by hour}
\usage{
birdnet_filter_hour(data, hour_arg)
}
\arguments{
\item{data}{A data frame with (or without) an \code{hour} column. If missing, datetime columns will be added.}

\item{hour_arg}{Integer vector of hours (0–23) to keep.}
}
\value{
A filtered data frame.
}
\description{
Filters BirdNET data to only include rows from specific hours of the day.
}
\keyword{internal}
