% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_row_editing.R
\name{birdnet_filter_year}
\alias{birdnet_filter_year}
\title{Filter BirdNET data by year}
\usage{
birdnet_filter_year(data, year_arg)
}
\arguments{
\item{data}{A data frame containing BirdNET output. If the \code{year} column is absent,
the function will attempt to extract datetime information using \link{birdnet_add_datetime}.}

\item{year_arg}{Integer vector. The year or years to retain in the filtered data.}
}
\value{
A filtered data frame containing only rows with detections from the specified years.
}
\description{
Filters a BirdNET output data frame based on one or more specified years of detection.
If the \code{year} column is missing, the function will extract it from available datetime columns.
}
\seealso{
\link{birdnet_add_datetime}, \link{birdnet_drop_datetime}
}
\keyword{internal}
