% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnetTools-package.R
\docType{package}
\name{birdnetTools-package}
\alias{birdnetTools}
\alias{birdnetTools-package}
\title{birdnetTools: Post-process, Validate, and Visualize BirdNET Output}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://birdnet-team.github.io/birdnetTools/}
  \item \url{https://github.com/birdnet-team/birdnetTools}
  \item Report bugs at \url{https://github.com/birdnet-team/birdnetTools/issues}
}

}
\author{
\strong{Maintainer}: Sunny Tseng \email{sunnyyctseng@gmail.com} (\href{https://orcid.org/0000-0002-8621-2244}{ORCID})

}
\keyword{internal}
