% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_row_editing.R
\name{birdnet_filter_date_range}
\alias{birdnet_filter_date_range}
\title{Filter data by date range}
\usage{
birdnet_filter_date_range(data, min_date = NULL, max_date = NULL)
}
\arguments{
\item{data}{A data frame with (or without) a \code{date} column. If missing, datetime columns will be added.}

\item{min_date}{Date or character string coercible to Date. The minimum date (inclusive).}

\item{max_date}{Date or character string coercible to Date. The maximum date (inclusive).}
}
\value{
A filtered data frame.
}
\description{
Filters BirdNET data to only include rows within a specified date range.
}
