% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_row_editing.R
\name{birdnet_filter_hour}
\alias{birdnet_filter_hour}
\title{Filter BirdNET data by hour of day}
\usage{
birdnet_filter_hour(data, hour_arg)
}
\arguments{
\item{data}{A data frame containing BirdNET output. If the \code{hour} column is not present,
the function will attempt to extract it using \link{birdnet_add_datetime}.}

\item{hour_arg}{Integer vector of hours (from 0 to 23) to retain.}
}
\value{
A data frame filtered to include only detections within the specified hours.
}
\description{
Filters a BirdNET output data frame to include only rows with detections
occurring during specified hours of the day (0–23). If the \code{hour} column
is missing, it will be extracted from available datetime columns.
}
\seealso{
\link{birdnet_add_datetime}, \link{birdnet_drop_datetime}
}
\keyword{internal}
