% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_row_editing.R
\name{birdnet_filter_species}
\alias{birdnet_filter_species}
\title{Filter data by species}
\usage{
birdnet_filter_species(data, species)
}
\arguments{
\item{data}{A data frame containing BirdNET output with a \code{common_name} column.}

\item{species}{Character vector. One or more common names of species to keep.}
}
\value{
A filtered data frame.
}
\description{
Filters BirdNET data for one or more specified species.
}
