% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_row_editing.R
\name{birdnet_filter_year}
\alias{birdnet_filter_year}
\title{Filter data by year}
\usage{
birdnet_filter_year(data, year_arg)
}
\arguments{
\item{data}{A data frame with (or without) a \code{year} column. If missing, datetime columns will be added.}

\item{year_arg}{Integer vector. The year(s) to keep.}
}
\value{
A filtered data frame.
}
\description{
Filters BirdNET data based on the year of detection.
}
