% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_subsample.R
\name{birdnet_subsample}
\alias{birdnet_subsample}
\title{Subsample BirdNET detections by species}
\usage{
birdnet_subsample(
  data,
  n,
  method = c("stratified", "random", "top"),
  save_to_file = FALSE,
  file = NULL
)
}
\arguments{
\item{data}{A data frame containing BirdNET output. It must include at least the columns \code{common_name} and \code{confidence}.}

\item{n}{Integer. Total number of observations to subsample for each species in \code{data}.}

\item{method}{Character string. Subsampling method to use. One of \code{"stratified"}, \code{"random"}, or \code{"top"}:
\describe{
\item{\code{"stratified"}}{Samples across confidence score strata (0.1 to 1 by 0.05 bins) evenly.}
\item{\code{"random"}}{Randomly samples \code{n} observations.}
\item{\code{"top"}}{Selects the top \code{n} observations with the highest confidence.}
}}

\item{save_to_file}{Logical. If \code{TRUE}, saves the output data frame to a file named \code{"subsampled_data.csv"} in the working directory. Default is \code{FALSE}. Automatically set to \code{TRUE} if \code{path} is defined.}

\item{file}{Character string or \code{NULL}. File path to save the output.
If \code{NULL} and \code{save_to_file = TRUE}, the file is saved as "subsampled_data.csv"
in the working directory.}
}
\value{
A data frame containing the subsampled observations.
}
\description{
Subsamples a specified number of observations for a given species from a BirdNET output dataset
using one of three methods: stratified, random, or top confidence. Optionally saves the result to a CSV file.
}
\examples{
\dontrun{
birdnet_subsample(data = my_data, species = "American Robin", n = 300, method = "stratified")
}
}
